rename ValueObjects to Values (removing the word Object)
