+++ app/Containers/Localization/Tasks/GetAllLocalizationsTask.php
@@ -2,7 +2,7 @@
 
 namespace App\Containers\Localization\Tasks;
 
-use App\Containers\Localization\ValueObjects\Localization;
+use App\Containers\Localization\Values\Localization;
 use App\Ship\Parents\Tasks\Task;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Config;
